+++ src/main/java/leets/enhance/domain/Blade/dto/response/RegisterResponse.java
@@ -1,4 +1,4 @@
-package leets.enhance.domain.Blade.response;
+package leets.enhance.domain.Blade.dto.response;
 
 public record RegisterResponse(Boolean result,
                                String message) {
