+++ src/main/java/leets/enhance/domain/Blade/dto/response/Top10ItemResponse.java
@@ -1,4 +1,4 @@
-package leets.enhance.domain.Blade.response;
+package leets.enhance.domain.Blade.dto.response;
 
 import java.util.List;
